% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{genheatmap}
\alias{genheatmap}
\title{Generate Heatmap

This function plots the heatmap of mean marker data for different cell types from the
training data.}
\usage{
genheatmap(df, is.log = FALSE)
}
\arguments{
\item{df}{Matrix of marker values (columns are different markers and rows are the cells)}

\item{is.log}{logical value denoting the need of log transformation.}
}
\value{
A heatmap plot
}
\description{
Generate Heatmap

This function plots the heatmap of mean marker data for different cell types from the
training data.
}
