% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_generator.R
\name{ridge_plottter}
\alias{ridge_plottter}
\title{Generate Ridge plots

This function generates ridge plot of a particular marker acroos different type of cells
for different images of the training datset}
\usage{
ridge_plottter(train_data, interested_trait, is.log = "TRUE")
}
\arguments{
\item{train_data}{the training dataset}

\item{interested_trait}{name of the marker of interest}

\item{is.log}{logical value, if TRUE, log transformed marker data will be considered}
}
\value{
A matrix of the infile
}
\description{
Generate Ridge plots

This function generates ridge plot of a particular marker acroos different type of cells
for different images of the training datset
}
