% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function_simple.R
\name{test_function_simple}
\alias{test_function_simple}
\title{Use the fitted random forest model simply on the test dataset

This function predicts the cell types of the test dataset which must have same number of
covariates as the train dataset.}
\usage{
test_function_simple(test_data, RFclassi, is.log = FALSE)
}
\arguments{
\item{test_data}{the input test dataset}

\item{RFclassi}{the fitted random forest model}

\item{is.log}{logical value denoting the need of log transformation. Should only be TRUE
if the train_function also have it as TRUE.}
}
\value{
A data.frame with all the cells of the test dataset along with their predicted classes
}
\description{
Use the fitted random forest model simply on the test dataset

This function predicts the cell types of the test dataset which must have same number of
covariates as the train dataset.
}
