% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_forest.R
\name{training_part}
\alias{training_part}
\title{Load a Matrix

This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.}
\usage{
training_part(train_data, is.log = FALSE)
}
\arguments{
\item{train_data}{to the input file}

\item{is.log}{to the input file}
}
\value{
A matrix of the infile
}
\description{
Load a Matrix

This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
