% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{summ_cell}
\alias{summ_cell}
\title{Summary of cell proportions}
\usage{
summ_cell(cell_data)
}
\arguments{
\item{cell_data}{The training dataset}
}
\value{
A table of cell counts of different types
}
\description{
This function provides a table of cell counts of different types in
the training dataset.
}
