% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_forest.R, R/training_part.R
\name{training_part}
\alias{training_part}
\title{Fits a random forest model on the train dataset

This function fits a random forest model based on the training dataset.}
\usage{
training_part(train_data, is.log = FALSE)

training_part(train_data, is.log = FALSE)
}
\arguments{
\item{train_data}{is the training dataset}

\item{is.log}{logical value denoting the need of log transformation. If its kept at TRUE, the same should be done
in test_function}
}
\value{
A matrix of the infile

A matrix of the infile
}
\description{
Fits a random forest model on the train dataset

This function fits a random forest model based on the training dataset.

Fits a random forest model on the train dataset

This function fits a random forest model based on the training dataset.
}
